+++ libs/shared/util/src/lib/object/index.ts
@@ -1,2 +1,2 @@
-export * from './nameOf/nameOf';
+export * from './of/of';
 export * from './deepFind/deepFind';
