+++ libs/shared/util/src/lib/object/of/of.spec.ts
@@ -1,4 +1,4 @@
-import { propnameOf, nameOf } from './nameOf';
+import { propnameOf, nameOf } from './of';
 
 const testObject = {
   someProperty: 'someValue',
