+++ libs/shared/util/src/lib/object/of/of.ts
